refactor: move swift2019 breadcrumbs into directory
