+++ src/components/pages-public/swift2019/bracketPlay/styles.js
