+++ src/components/pages-public/swift2019/bracketPlay/tableLabel/index.js
