+++ src/components/pages-public/swift2019/bracketPlay/tableLabel/styles.js
